Allow Python children to return strings interpreted as subqueries.
These subqueries are assumed to return result sets of the same shape and types
as the original query. They are executed via the SPI and sent directly back to
the foreign scan routine, avoiding costly Python type conversions.

It's possible to mix and match items in the returned iterator: for example,
if first a string is returned and then a tuple, Multicorn will first return
results from the subquery and then treat the tuple as a normal row (so
backwards compatibility is preserved).

Also changed the API to send columns/quals as lists instead of sets:
otherwise, the ordering of columns isn't preserved and the Python
execute() method will return a query with an unexpected permutation of columns.