Fix a bug where LQs with a subset of all columns would return garbage.
This was due to the TupleTableSlot given to Multicorn sometimes having
extra columns than what Postgres requested (eg the whole table) and
Multicorn would insert the tuples returned by the subquery in the beginning
of the TupleTableSlot (causing them to be interpreted incorrectly).

For the fix, go through the TupleTableSlot and match up (crudely: if
the subquery returned columns in a different order than they're listed
in the TTS, this will fail) the columns we requested from Python with
the columns in the slot and insert them into the right positions.