Enable aggregation pushdown for columns other than first
Given the mechanism of PG FDW aggregation pushdown, whereby the target list does not have any references to the original table,
the existing initConversioninfo function was modified, so as to match the tuple table of the aggregation scan state. This was
done by simply ommiting cinfo records that do not appear in the target_list.