Implmentation of aggregation/grouping pushdown
The current implementation provides a mechanism for pushing down aggregation and/or grouping
queries into the foreign data source. The Python side of the implementation will now receive
two new kwargs, `aggs` and `group_clauses`, in which case it should return the corresponding
aggreagation result.

Still left to implement is consulting the Python side whether remote aggregation is possible
at all, and if so which agregation functions are valid.

Also missing are some more advanced aggregation cases (aggregating multiple functions, or
handling `HAVING` clause for example). This is to be implemented separately.